% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuarFuns.R
\name{actuar_qtrbeta}
\alias{actuar_qtrbeta}
\title{TrB Quantile}
\usage{
actuar_qtrbeta(p, abcd)
}
\arguments{
\item{p}{A numeric vector of cumulative probabilities.}

\item{abcd}{A numeric matrix of 4 rows. Each column is a vector of 
\code{a, b, c, d}.}
}
\value{
A numeric vector of the quantiles, \code{q}. \code{q[i]} is the 
quantile corresponding to \code{p[i]} for TrB parameterized by 
\code{abcd[, i]}.
}
\description{
Inverse TrB CDF of \code{a, b, c, d}.
}
\examples{
p = runif(100)
abcd = matrix(runif(4 * 100, 1, 10), ncol = 100)
NGFMfitDistr::actuar_qtrbeta(p, abcd)
}
