% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWindows.R
\name{computeWindows}
\alias{computeWindows}
\title{Compute sliding windows}
\usage{
computeWindows(
  xlen,
  windowSize = 1L,
  start = 1L,
  speed = 1L,
  returnList = TRUE
)
}
\arguments{
\item{xlen}{Size of the full data.}

\item{windowSize}{A positive integer as the window size. Default to 1.}

\item{start}{A positive integer. Index of the 1st data point included in
the 1st window.}

\item{speed}{A positive integer. Sliding speed.}

\item{returnList}{A Boolean value. If \code{TRUE}, the windows are 
returned as a list of integer vectors. Each integer vector is of size 2. 
The 1st element is the index of the first data point in window, the 2nd the
last data point in window. If \code{FALSE}, the list is converted to a 
2-row matrix.}
}
\value{
If \code{returnList} is \code{TRUE}, a list of integer vectors of 
size 2. Otherwise an integer matrix with 2 rows.
}
\description{
Compute sliding windows given window size and sliding speed.
}
\details{
More details can be found in the slide documentation 
\href{../doc/slides.pdf}{\code{vignette('slides', package = 'NGFMfitDistr')}}.
}
\examples{
NGFMfitDistr::computeWindows(
  xlen = 30L, windowSize = 7L, start = 1L, speed = 3L, returnList = TRUE)


NGFMfitDistr::computeWindows(
  xlen = 30L, windowSize = 7L, start = 1L, speed = 3L, returnList = FALSE)
}
