% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractMain}
\alias{extractMain}
\title{Extract main part.}
\usage{
extractMain(distlist, MDR, normalizeMainPart = TRUE, removeZeroPs = FALSE)
}
\arguments{
\item{distlist}{A list of PMFs. Each PMF is a list of two numeric vectors.
The first is the support. The second is the probabilities.}

\item{MDR}{A numeric vector of the prescribed MDRs. Should have the same
size as \code{distlist}.}

\item{normalizeMainPart}{A boolean value. Should we normalize the 
probabilities after removing \eqn{P_0}? Default to \code{TRUE}.}

\item{removeZeroPs}{A boolean value. Some probabilities can be zeros in
\code{distlist}. Should we remove them along with their points on the
support? Default to \code{FALSE}.}
}
\value{
A list of two:

\code{$conditionalDistrs}:  A list of PMFs to be fitted by TrBs.

\code{$lm1}: A numeric vector of the target limited means 
for constraining \eqn{\mathbb{E}[\min(X, 1)]} of the TrB models.
}
\description{
Remove \eqn{P_0}s from the PMFs in a list.
}
\details{
More details can be found in the slide documentation 
\href{../doc/slides.pdf}{\code{vignette('slides', package = 'NGFMfitDistr')}}.
}
\examples{
distlist = lapply(1:10, function(x) list(val = sort(runif(10)), P = runif(10)))
for (i in 1:length(distlist)) 
  distlist[[i]]$P = distlist[[i]]$P / sum(distlist[[i]]$P)
MDR = runif(10)
rst = extractMain(distlist, MDR, normalizeMainPart = TRUE,
            removeZeroPs = FALSE)
}
