% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullDiscretize.R
\name{fullDiscretize}
\alias{fullDiscretize}
\title{Full Pipe for Inflated Transformed Beta Discretization}
\usage{
fullDiscretize(
  TrBtable,
  supportSizes = c(42L, 64L),
  regridMethod = "lr",
  RIBlib = "Numerical Recipes",
  outputProbsInRows = FALSE,
  fineDiscretizationSize = 2000,
  maxCore = 1000,
  verbose = TRUE,
  downScaleSupport = FALSE,
  figureDir = "../figure",
  empDistrsLists = NULL,
  mdrRangeInData = NULL,
  claimsDataForScoring = NULL
)
}
\arguments{
\item{TrBtable}{A numeric matrix of 7 rows. Each column is a vector of
\code{(MDR, max, }\eqn{P_0}\code{, a, b, c, d)}.}

\item{supportSizes}{An integer vector, e.g. \code{c(64, 42)}.}

\item{regridMethod}{Regrid method: \code{"lr"} (linear regrid) or 
\code{"lmm"} (local moment matching) or \code{"r4"} (four-point regrid).
Default to \code{"lr"}.}

\item{RIBlib}{A string to specify which library should be used to 
compute the \href{https://en.wikipedia.org/wiki/Beta_function#Incomplete_beta_function}{Regularized Incomplete Beta function}
that TrB CDF depends on. 
\itemize{
\item{\code{"Numerical Recipes"}}{ implements the textbook algorithm in
\href{https://e-maxx.ru/bookz/files/numerical_recipes.pdf}{Chapter 6.4, Numerical Recipes}.
\href{https://live.boost.org/doc/libs/1_54_0/libs/math/doc/html/math_toolkit/sf_beta/ibeta_function.html}{Boost}, 
which claims the textbook algorithm's series expansion converges less 
rapidly than its own approach, actually runs considerably slower because 
it pursues much higher numeric precision.
Such extreme precision turns out largely unncessary in our setting. 
\code{"Numerical Recipes"} is the default and 
the recommended choice.}

\item{\code{"R::pbeta"}}{ calls R's Beta CDF. It is slower than \code{"Numerical Recipes"}
but faster than \code{"Boost"}. In numerically pathological regions, 
\code{"R::pbeta"} is more precise than \code{"Numerical Recipes"} 
but less so than \code{"Boost"}.}

\item{\code{"Boost"}}{ calls \href{https://live.boost.org/doc/libs/1_54_0/libs/math/doc/html/math_toolkit/sf_beta/ibeta_function.html}{boost::math::ibeta}.
It is the slowest but yields much higher precision especially 
in numerically pathological regions.}
}}

\item{outputProbsInRows}{A boolean. \code{TRUE} will produce a PMF 
table (matrix) where every row is a vector of (\code{MDR, max}, \eqn{P_0}, 
\code{P1, ..., Pmax}). \code{FALSE} will store it in a column. 
Default to \code{FALSE}.}

\item{fineDiscretizationSize}{An integer. Support size for raw 
discretization before tunning. Default to 2000.}

\item{maxCore}{The number of CPU cores for use. Default to 1000. A number
greater than the actual number of logical processors will be capped.
The bidirectional sequential fitting, which is the most likely to be used, 
is single-threaded if the initialization \code{abc} is singular --- when
\eqn{K=1}.}

\item{verbose}{A boolean. \code{TRUE} prints QA checking messages.}

\item{downScaleSupport}{A boolean. When the raw discretization's mean is
less than MDR, \code{TRUE} will downscale the support to match the MDR, 
while \code{FALSE} will increase \eqn{P_0} and downscale the main part 
probabilities to match the MDR. Default to \code{FALSE}, which is the 
better choice from experiments.}

\item{figureDir}{Path to the directory for saving plots. 
\code{figureDir = ""} silences visualization. The directory will be created
if nonexistent. Default to \code{"../figure"}.}

\item{empDistrsLists}{The list of empirical PMFs to which inflated TrBs 
are fitted. \code{empDistrsLists} is a member in the output from \code{\link{fullFit}}. 
The object is only needed for visualization, and would be ignored if
\code{figureDir=""}. Default to \code{NULL} which silences 
visualization.}

\item{mdrRangeInData}{A numeric vector of size 2. 
MDR range in the original claims data. \code{mdrRangeInData}
is a member in the output from \code{\link{fullFit}}. The object is only needed 
for visualization, and would be ignored if \code{figureDir=""}. Default to
\code{NULL} which silences visualization.}

\item{claimsDataForScoring}{Claims data for scoring the tuned PMF table.
See argument \code{claimsData} in \link{scorePMFtable} for more
details. Default to \code{NULL} which also silences scoring.}
}
\value{
A list of the results from \code{\link{FTDFQA}} and the results from 
\code{\link{scorePMFtable}}. The latter exists only if claims data are given to 
score the PMF table.
}
\description{
Discretize inflated TrB distributions characterized by a parameter table 
under QA requirements. This function amalgamates \code{\link{FTDFQA}}, 
\code{\link{plotQAparams}}, \code{\link{plot45pmfsInTable}}, 
\code{\link{plot207pmfsInTable}} and \code{\link{scorePMFtable}}.
}
