% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{longestNonDecreasingSubseq}
\alias{longestNonDecreasingSubseq}
\title{Longest Nondecreasing Subsequence.}
\usage{
longestNonDecreasingSubseq(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An integer vector of the 1-based indices of the elements in the 
subsequence of interest.
}
\description{
Compute the longest nondecreasing subsequence from a given sequence of numbers.
}
\details{
More details can be found \href{https://en.wikipedia.org/wiki/Longest_increasing_subsequence}{here}.
}
\examples{
set.seed(123)
x = runif(300)
y = NGFMfitDistr::longestNonDecreasingSubseq(x)
plot(x, col = "blue")
lines(y, col = "red")
}
