% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMax.R
\name{makeMax}
\alias{makeMax}
\title{Make Maxes}
\usage{
makeMax(
  MDR,
  P0,
  abcd,
  tailThreshold = 0.01,
  minMax = 0.05,
  interpolationMethod = "linear"
)
}
\arguments{
\item{MDR}{A numeric vector of MDRs.}

\item{P0}{A numeric vector of \eqn{P_0}s. Should have the same length of \code{MDR}.}

\item{abcd}{A \eqn{4 \times} \code{length(MDR)} numeric matrix of TrB 
parameters.}

\item{tailThreshold}{A tail probability for computing a tentative max.
This max is chosen such that the total probability of damage ratio above 
the max will not exceed \code{tailThreshold}.}

\item{minMax}{Max for the lowest MDR. Default to 0.05.}

\item{interpolationMethod}{Interpolation method, either \code{"linear"} or 
\code{"hyman"}. Default to \code{"linear"}, which invokes 
\code{\link[stats]{approxfun}(method = "linear")}.
\code{"hyman"} invokes \code{\link[stats]{splinefun}(method = "hyman")}. 
The argument controls how we interpolate the longest monotonic subsequence
of \eqn{P_0} (or \eqn{\max}). The interpolation function is then used to
estimate the \eqn{P_0} (or \eqn{\max}) at any MDR.}
}
\value{
A numeric vector of maxes corresponding to \code{MDR}.
}
\description{
Estimate proper maxes of the supports for TrB discretization.
}
\details{
The function (i) computes a sequence of tentative maxes for 
the given \code{MDR}s, (ii) searches the longest nondecreasing subsequence,
(iii) interpolates the subsequence and estimates all the maxes for \code{MDR}.
}
