% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{plot207pmfsInTable}
\alias{plot207pmfsInTable}
\title{Plot 207 PMFs in PMF table}
\usage{
plot207pmfsInTable(figPath, pmfTable, empDistrsLists, mdrRangeInData)
}
\arguments{
\item{figPath}{Path to the figure to be saved. File extension MUST be
\code{.pdf}.}

\item{pmfTable}{The distribution table as a matrix. Every column is 
\code{MDR, max, P0, P1, ..., Pmax}.}

\item{empDistrsLists}{Result from \code{\link{estimateEmpPMFs}}.}

\item{mdrRangeInData}{A numeric vector of size 2. Range of MDRs whose 
associated empirical PMFs are inferred from data.}
}
\value{
None
}
\description{
Plot PMFs associated with MDR = 0.00001, 0.00002, ..., 0.0001, 0.0002, ...
0.001, 0.002, ..., 0.01, 0.02, ..., 0.1, 0.2, ..., 0.9.
}
\details{
A helper function for visualization. Please walk 
through \code{vignette('Workflow', package = 'NGFMfitDistr')} to see how it 
is used.
}
