% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{plotP0model}
\alias{plotP0model}
\title{Plot \eqn{P_0}}
\usage{
plotP0model(figPath, targetMDRs, p0models)
}
\arguments{
\item{figPath}{Path to the figure to be saved. File extension MUST be
\code{.png} or \code{.PNG}.}

\item{targetMDRs}{A numeric vector of MDRs.}

\item{p0models}{A list returned from function \code{\link{ensembleP0}}.}
}
\value{
None
}
\description{
Plot fitted \eqn{P_0} versus MDRs.
}
\details{
A helper function for visualization. Please walk 
through \code{vignette('Workflow', package = 'NGFMfitDistr')} to see how it 
is used.
}
