% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{plotQAparams}
\alias{plotQAparams}
\title{Visualize Parameter Trends in PMF table}
\usage{
plotQAparams(pmfTable, figPath)
}
\arguments{
\item{pmfTable}{PMF table as a numeric matrix. Every
column is a (MDR, max, P0, P1, ..., Pmax). The first and last PMFs will be 
excluded from visualization.}

\item{figPath}{Path to the figure to be saved. File extension must be
\code{.png} or \code{.PNG}.}
}
\value{
None
}
\description{
Plot parameter trends in QA's interest, namely \eqn{P_0}s, \eqn{P_{\max}}s, 
maxes, PMF means, coefficients of variation against MDR in a PMF table.
}
