% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{plotTrBparameters}
\alias{plotTrBparameters}
\title{Plot TrB Parameters}
\usage{
plotTrBparameters(figPath, optRst, startingPmf, abc)
}
\arguments{
\item{figPath}{Path to the figure to be saved. File extension MUST be
\code{.png} or \code{.PNG}.}

\item{optRst}{Optimization result from function \code{\link{LBFGSBtrbFitList}}.}

\item{startingPmf}{An integer as the index of the first PMF that was fitted
during bi-directional sequential fitting.}

\item{abc}{A numeric vector of size 3. The initial \code{a, b, c} for fitting
the \code{startingPmf}-th empirical PMF.}
}
\value{
None
}
\description{
Plot trained Transformed Beta parameters versus MDRs.
}
\details{
A helper function for visualization. Please walk 
through \code{vignette('Workflow', package = 'NGFMfitDistr')} to see how it 
is used.
}
