% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_d}
\alias{solve_d}
\title{Solve the scale parameter d.}
\usage{
solve_d(
  abc_lm1,
  eps = 1e-08,
  maxit = 100L,
  useNewton = TRUE,
  RIBlib = "R::pbeta"
)
}
\arguments{
\item{abc_lm1}{A numeric matrix with four rows. Each column is a vector of
\code{(a, b, c)} and the target limited mean at 1.}

\item{eps}{If the difference between TrB's limited mean and \code{lm1} is
less than \code{eps}, stop the Newton's iteration. Default to \code{1e-8}.}

\item{maxit}{The maximum number of iterations for Newton's method.
Default to \code{100}.}

\item{useNewton}{A boolean. \code{TRUE} invokes Newton's method. 
\code{FALSE} invokes bisection. Default to \code{TRUE}.}

\item{RIBlib}{A string to specify which library should be used to 
compute the \href{https://en.wikipedia.org/wiki/Beta_function#Incomplete_beta_function}{Regularized Incomplete Beta function}
that TrB CDF depends on. 
\itemize{
\item{\code{"Numerical Recipes"}}{ implements the textbook algorithm in
\href{https://e-maxx.ru/bookz/files/numerical_recipes.pdf}{Chapter 6.4, Numerical Recipes}.
\href{https://live.boost.org/doc/libs/1_54_0/libs/math/doc/html/math_toolkit/sf_beta/ibeta_function.html}{Boost}, 
which claims the textbook algorithm's series expansion converges less 
rapidly than its own approach, actually runs considerably slower because 
it pursues much higher numeric precision.
Such extreme precision turns out largely unncessary in our setting. 
\code{"Numerical Recipes"} is the default and 
the recommended choice.}

\item{\code{"R::pbeta"}}{ calls R's Beta CDF. It is slower than \code{"Numerical Recipes"}
but faster than \code{"Boost"}. In numerically pathological regions, 
\code{"R::pbeta"} is more precise than \code{"Numerical Recipes"} 
but less so than \code{"Boost"}.}

\item{\code{"Boost"}}{ calls \href{https://live.boost.org/doc/libs/1_54_0/libs/math/doc/html/math_toolkit/sf_beta/ibeta_function.html}{boost::math::ibeta}.
It is the slowest but yields much higher precision especially 
in numerically pathological regions.}
}}
}
\value{
A numeric vector of \code{d}s. The vector is of of size \code{ncol(abc_lm1)}.
}
\description{
Solve TrB parameter d given a, b, c and the constrained limited mean at 1.
}
